Clean up motor class
Use variables for max and min powers.

Remove pin format since it is checked at a lower level anyway.

Instead of logging and not setting power when it is out of bounds,
simply cap it to the bound.

Remove the destructor. There is no guarantee that it will be called
it is best to simply use PWM.cleanup()