control: Reduce the size of FIFO buffer before reset
The MPU 6050 has a FIFO buffer that is 1024 bytes long. As the buffer
approaches that size limit, it starts giving bad values.

One method is to increase the rate at which it reads. The other is to
reduce the buffer size. This commit addresses the latter.

Also, a smaller buffer size (in this case, of 10 packets size)
ensures fresher values.