getMpuAngles(): Reset FIFO if buffer is greater than desired size
The buffer was reset only if fifoCount was equal to the desired
number. If, by some chance, this condition was not checked when
the size of the buffer was exactly as required, it would never
reset an overflowing buffer.

Therefore, reset the buffer when the value exceeds the limit but
only when an entire packet is available since resetting when a
partial packet is available messes up the buffer.