MPU6050: Increase FIFO rate to 100Hz
The FIFO rate was 40Hz. That is a value every 25ms. Faster orientation
values can yield better performance.

Setting the divisor to 0x02 gives you (200 / 2) = 100Hz. That is, a
value every 10ms.

To keep up with that pace, the clock frequency of the I2C bus on the
BeagleBone Black had to be increased to be 400KHz, up from the 100KHz.

This can be changed on boot up by decompiling the dtb file at
/boot/dtbs/3.8.13-bone70/am335x-boneblack.dtb and changing the clock
cycle, before recompiling and rebooting.

To keep the values fresh on the buffer, the buffer was also made to
reset if the queue length reaches 3 packets.