PID: Move and correct error calculation
The error calculation algorithm written earlier to correct when angles
from the IMU changed from +/- 179 to -/+ 179 had a flaw when the
desired angle was positive and the actual angle negative.

Eg:
desired = -170, actual = 170
error = desired - actual = -170 - 170 = -340

error < 360 - error is true and error remains -340.

The correct error should be +20.

----

The error calculation has also been moved back into the PID class
since it does not make sense to pass in both the error and the
feedback_value.

Co-authored-by: Karthik Nayak <Karthik.188@gmail.com>