Add parse and execute function
parse_and_execute() parses the control_string string for a
command and a value that is separated by a space.

The first token must be a command and the second a value that
can be converted into a float value. It should not contain a
newline character.

The parsed data is then updated for use by the PID.

TODO: Pass in control_string instead of having a global variable.